backward compatibility fix
